// src/api/schemas/userConfigKey.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * MyApp API
 * OpenAPI spec version: 1.0.0
 */

/**
 * Enum for user configuration keys.
This is used in the OpenAPI schema for type safety.
 */
export type UserConfigKey = (typeof UserConfigKey)[keyof typeof UserConfigKey];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserConfigKey = {
  enable_email_notifications: 'enable_email_notifications',
  enable_sound_on_discussion_notification: 'enable_sound_on_discussion_notification',
} as const;