// src/api/schemas/userSettingsResponseSchema.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * MyApp API
 * OpenAPI spec version: 1.0.0
 */
import type { UserSettingSchema } from './userSettingSchema';

/**
 * Schema for getting all user settings.
 */
export interface UserSettingsResponseSchema {
  settings: UserSettingSchema[];
}